Retrieve course user from db for response
Previously the `createSessionUserRecord` function was programmatically
updating a course user object and then serialising it for a response.

This method led to a bug where the course user did not have a session user
array yet but the function attempted to push a new session user to a
non-existent array.

This change retrieves the course user directly from the db.
The course user retrieved in this way will automatically contain
session user information.

140622-1802 / Rollbar: Failed POST "/api/v1/session-user"