Add db logic for cancelling a whatsapp subscription
If an active subscription exists, this method will cancel it by filling in the cancelledAt column.

Even though this is a PATCH method, it can only be used to cancel a subscription, never to activate. If used to activate, the method throws an error.

b1d98ae473 / Create whatsapp subscribe endpoint