Update therapy session table to handle null timezone
This will handle the case where simplybook does not send timezone information for a booking.
This happens when the therapist makes the booking on the client's behalf.
The client timezone is not available in this case.
It is always available when the client makes their own booking.

d0ee58e1b8 / Handle empty timezone in simplybook webhook